refactor: extract AbstractOAuth2Authenticator base class (#26)
Move all shared OAuth2 logic (constructor, callbacks, handleAuthorizationCode,
state validation, HTTP transport via cURL) into AbstractOAuth2Authenticator.
AzureAuthenticator and GoogleAuthenticator now extend the base and only contain
provider-specific code (token exchange, user-info retrieval, logon processing).

Also fixes:
- double curl_getinfo/unset bug in AzureAuthenticator::sendPost
- missing Kingsoft\Google and Kingsoft\OAuth2 namespaces in composer.json autoload
- global-namespace prefix on built-in functions (call_user_func, get_class)

Closes #26
